Refactor out reusable card-stack component
