+++ BlazarUI/app/scripts/components/shared/RepoBranchCard.jsx
@@ -1,293 +1,31 @@
-import React, { PropTypes } from 'react';
-import classNames from 'classnames';
-import { Link } from 'react-router';
-import moment from 'moment';
+import React, {Component} from 'react';
+import Card from '../shared/card-stack/Card.jsx';
 
-import Card from './Card.jsx';
-import ModuleRow from './ModuleRow.jsx';
-import BuildStates from '../../constants/BuildStates';
-import Loader from '../shared/Loader.jsx';
-import Sha from '../shared/Sha.jsx';
+import RepoBranchCardSummary from '../shared/RepoBranchCardSummary.jsx';
+import RepoBranchCardDetails from '../shared/RepoBranchCardDetails.jsx';
 
-class RepoBranchCard extends Card {
+function RepoBranchCard({item, moduleBuildsList, expanded, onClick}) {
+  const summary = (
+    <RepoBranchCardSummary
+      item={item}
+    />
+  );
 
-  getClassNames() {
-    const build = this.getBuildToDisplay();
-    const isFailingBuild = build.get('state') === BuildStates.FAILED
-      || build.get('state') === BuildStates.UNSTABLE;
+  const details = (
+    <RepoBranchCardDetails
+      item={item}
+      moduleBuildsList={moduleBuildsList}
+    />
+  );
 
-    return classNames([
-      this.getBaseClassNames(),
-      'card-stack__card--repo-branch', {
-        'repo-branch-card--failed': isFailingBuild
-      }
-    ]);
-  }
-
-  sortModules() {
-    const {moduleBuildsList} = this.props;
-    const build = this.getBuildToDisplay().toJS();
-
-    if (!build || !build.dependencyGraph || !build.dependencyGraph.topologicalSort) {
-      return moduleBuildsList;
-    }
-
-    const topologicalSort = build.dependencyGraph.topologicalSort;
-
-    return moduleBuildsList.sort((a, b) => {
-      const indexA = topologicalSort.indexOf(a.moduleId);
-      const indexB = topologicalSort.indexOf(b.moduleId);
-
-      if (indexA < indexB) {
-        return -1;
-      }
-
-      return 1;
-    });
-  }
-
-  getBuildTime(startTimestamp) {
-    const timestampText = moment(startTimestamp).fromNow();
-
-    if (timestampText === 'a day ago') {
-      return 'yesterday';
-    }
-
-    return timestampText;
-  }
-
-  getBuildToDisplay() {
-    const {item} = this.props;
-
-    if (item.get('inProgressBuild') !== undefined) {
-      return item.get('inProgressBuild');
-    }
-
-    return item.get('lastBuild');
-  }
-
-  renderRepoLink() {
-    const {item} = this.props;
-    const gitInfo = item.get('gitInfo');
-
-    return (
-      <Link to={gitInfo.get('blazarRepositoryPath')}>
-        {gitInfo.get('repository')}
-      </Link>
-    );
-  }
-
-  renderBuildNumberLink() {
-    const build = this.getBuildToDisplay();
-
-    return (
-      <Link className="repo-branch-card__build-number" to={build.get('blazarPath')}>
-        #{build.get('buildNumber')}
-      </Link>
-    );
-  }
-
-  buildCompareLink() {
-    const commitInfo = this.getBuildToDisplay().get('commitInfo');
-    const previousCommit = commitInfo.get('previous');
-
-    if (!previousCommit) {
-      return null;
-    }
-
-    const currentCommit = commitInfo.get('current');
-    const commitUrl = `${previousCommit.get('url').replace('/commit/', '/compare/')}...${currentCommit.get('id')}`;
-
-    return (
-      <a href={commitUrl}>
-        compare
-      </a>
-    );
-  }
-
-  renderModuleRows(modules) {
-    return modules.map((module, i) => <ModuleRow module={module} key={i} />);
-  }
-
-  renderDetails() {
-    const {moduleBuildsList, expanded, loading} = this.props;
-    const build = this.getBuildToDisplay();
-
-    if (!expanded) {
-      return (
-        <div className="card-stack__expanded collapsed">
-          <div className="card-stack__expanded-module-rows collapsed" />
-        </div>
-      );
-    } else if (loading) {
-      return (
-        <div className="card-stack__expanded">
-          <Loader align="center" />
-        </div>
-      );
-    }
-
-    let buildTriggerMessage;
-    let detailedTriggerMessage;
-
-    if (build.get('buildTrigger') && build.get('buildTrigger').get('type') === 'MANUAL') {
-      buildTriggerMessage = 'manually';
-      const buildTime = this.getBuildTime(build.get('startTimestamp'));
-      const author = build.get('buildTrigger').get('id');
-      detailedTriggerMessage = `Triggered ${buildTime}${author === 'unknown' ? '' : ` by ${author}`}`;
-    } else {
-      buildTriggerMessage = 'automatically by a code push';
-      detailedTriggerMessage = this.buildCompareLink();
-    }
-
-    return (
-      <div className="card-stack__expanded">
-        <div className="repo-branch-card__expanded-header">
-          <span>Build {this.renderBuildNumberLink()} was started {buildTriggerMessage}</span>
-          <span className="repo-branch-card__expanded-author">{detailedTriggerMessage}</span>
-        </div>
-        <div className="repo-branch-card__expanded-module-rows">
-          {this.renderModuleRows(this.sortModules(moduleBuildsList))}
-        </div>
-      </div>
-    );
-  }
-
-  renderInfo() {
-    const {item} = this.props;
-    const gitInfo = item.get('gitInfo');
-
-    return (
-      <div className="repo-branch-card__info">
-        <div className="repo-branch-card__repo">
-          {this.renderRepoLink()}
-        </div>
-        <div className="repo-branch-card__branch-and-build">
-          <span className="repo-branch-card__branch">
-            <Link to={gitInfo.get('blazarBranchPath')}>
-              {gitInfo.get('branch')}
-            </Link>
-          </span>
-        </div>
-      </div>
-    );
-  }
-
-  renderBuildAuthorMaybe() {
-    const trigger = this.getBuildToDisplay().get('buildTrigger');
-    const author = trigger.get('id');
-
-    if (trigger && trigger.get('type') !== 'MANUAL' || author === 'unknown') {
-      return null;
-    }
-
-    return (
-      <span>
-        Triggered by <span className="repo-branch-card__author">{author}</span>
-      </span>
-    );
-  }
-
-  renderLastBuild() {
-    const build = this.getBuildToDisplay();
-    let timestamp;
-
-    if (build.get('state') === BuildStates.IN_PROGRESS) {
-      timestamp = 'In Progress';
-    } else {
-      timestamp = this.getBuildTime(build.get('startTimestamp'));
-    }
-
-    return (
-      <div className="repo-branch-card__last-build">
-        <span className="repo-branch-card__last-build-time">
-          {timestamp}
-        </span>
-      </div>
-    );
-  }
-
-  renderTriggeredBy() {
-    const build = this.getBuildToDisplay();
-    let buildTriggerMessage;
-    let sha;
-
-    if (build.get('buildTrigger') && build.get('buildTrigger').get('type') === 'MANUAL') {
-      const buildAuthor = this.renderBuildAuthorMaybe();
-      if (!buildAuthor) {
-        buildTriggerMessage = 'Triggered by user';
-      } else {
-        buildTriggerMessage = buildAuthor;
-      }
-    } else {
-      buildTriggerMessage = 'Triggered by code push';
-      const commitInfo = build.get('commitInfo');
-
-      if (build.get('sha') && commitInfo) {
-        const gitInfo = {
-          host: commitInfo.get('host'),
-          organization: commitInfo.get('organization'),
-          repository: commitInfo.get('repository')
-        };
-
-        sha = (<Sha gitInfo={gitInfo} build={build.toJS()} />);
-      }
-    }
-
-    return (
-      <div className="repo-branch-card__details">
-        <div className="repo-branch-card__triggered-by">
-          <span>{buildTriggerMessage} {sha ? '(' : ''}{sha}{sha ? ')' : ''}</span>
-        </div>
-      </div>
-    );
-  }
-
-  renderStatus() {
-    const build = this.getBuildToDisplay();
-    const classWithModifier = `repo-branch-card__status-banner--${[BuildStates.FAILED, BuildStates.UNSTABLE].indexOf(build.get('state')) !== -1 ? 'FAILED' : 'SUCCESS'}`;
-
-    return (
-      <div className="repo-branch-card__status">
-        <div className={classWithModifier}>
-          <span />
-        </div>
-      </div>
-    );
-  }
-
-  renderBuildAndStatus() {
-    const build = this.getBuildToDisplay();
-    const colorClass = `repo-branch-card__build-and-status repo-branch-card__build-and-status--${build.get('state')}`;
-
-    return (
-      <div className={colorClass}>
-        {this.renderBuildNumberLink()}
-      </div>
-    );
-  }
-
-  render() {
-    return (
-      <div className={this.getClassNames()}>
-        <div className="repo-branch-card__inner-wrapper" onClick={this.props.onClick}>
-          {this.renderBuildAndStatus()}
-          <div className="card-stack__card-main">
-            {this.renderInfo()}
-            {this.renderLastBuild()}
-            {this.renderTriggeredBy()}
-          </div>
-        </div>
-        {this.renderDetails()}
-      </div>
-    );
-  }
+  return (
+    <Card
+      summary={summary}
+      details={expanded ? details : null}
+      expanded={expanded}
+      onClick={onClick}
+    />
+  );
 }
-
-RepoBranchCard.propTypes = {
-  item: PropTypes.object.isRequired,
-  loading: PropTypes.bool.isRequired,
-  moduleBuildsList: PropTypes.array
-};
 
 export default RepoBranchCard;
