--- BlazarUI/app/stylus/card-variables.styl
@@ -1,13 +1,0 @@
-$card-stack-border-color = #eaf0f6
-$card-stack-border-color-dark = #CBD6E2
-$card-bg-color = #ccedf2
-$card-bg-color-hover = rgba(229, 246, 248, 0.1)
-$border = 1px solid $card-stack-border-color
-$border-dark = 1px solid $card-stack-border-color-dark
-$initial-height = 80px
-$transition-duration = .3s
-$transition-timing-func = ease
-$bg-transition-duration = .15s
-$shadow-depth-1 = 0 3px 6px rgba(0,0,0,0.20)
-$shadow-depth-2 = 0 6px 6px rgba(0,0,0,0.20)
-$shadow-depth-3 = 0 10px 10px rgba(0,0,0,0.20);
