+++ BlazarUI/app/stylus/components/card-stack/card-variables.styl
@@ -1,0 +1,10 @@
+$border-color = #cbd6e2
+$border-width = 1px
+$expanded-card-margin = 30px
+
+$transition-duration = .3s
+$transition-timing-func = ease
+
+$shadow-depth-1 = 0 3px 6px rgba(0,0,0,0.20)
+$shadow-depth-2 = 0 6px 6px rgba(0,0,0,0.20)
+$shadow-depth-3 = 0 10px 10px rgba(0,0,0,0.20);
