+++ BlazarUI/app/stylus/main.styl
@@ -22,7 +22,7 @@
 @import 'components/build-log'
 @import 'components/build-log-navigation'
 @import 'components/branch-filter'
-@import 'components/card-stack'
+@import 'components/card-stack/card-stack'
 @import 'components/clickable-table-row'
 @import 'components/collapsable'
 @import 'components/commits'
