+++ BlazarUI/app/stylus/page/dashboard.styl
@@ -20,6 +20,10 @@
 
 .dashboard__toggle-container
   float right
-  
+
 .dashboard__section-title
   font-weight $font-normal
+
+.dashboard__repo-branch-card-stack
+  margin 60px 20px
+  min-width 800px
